+++ src/AI_BrowserFolders/AI_BrowserFolders.cs
@@ -1,4 +1,5 @@
-﻿using BepInEx;
+﻿using System.Diagnostics;
+using BepInEx;
 
 namespace BrowserFolders
 {
@@ -6,5 +7,14 @@ namespace BrowserFolders
     [BepInProcess("StudioNEOV2")]
     public partial class AI_BrowserFolders : BaseUnityPlugin
     {
+        [Conditional("itsempty")] // remove if there's any actual code added
+        private void GameSpecificAwake()
+        {
+        }
+
+        [Conditional("itsempty")]
+        private void GameSpecificOnGui()
+        {
+        }
     }
 }
