+++ src/HS2_BrowserFolders/HS2_BrowserFolders.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -81,6 +81,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="HS2_BrowserFolders.cs" />
+    <Compile Include="MainGameFolders.cs" />
+    <Compile Include="NestedFilenamesMainGamePatch.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
